Wrapped up with new vocab endpoint approach
